Create backupserverfilestos3.groovy
This pipeline will download the Backup files stored in the S3 bucket. It fetch the backup files of Nginx and Mysql with Date. Before running this Pipeline, run the shell script to download the backup files from EC2 to S3 bucket. Once the files are ready, go ahead and run this pipeline so you can download the backup files from s3 for the particular date. This script is prepared to download the files on a daily basis.