Create downloadfilefroms3.groovy
This pipeline will download the file from the s3 bucket. Before running this script, make sure the file is existing in the S3 bucket. A shell script should run before this pipeline, so the shell script move the file from EC2 to S3 bucket and we can download it from there using this pipeline. Also, it download the file based on the current date. So, make sure you're running this pipeline on a daily basis to get the daily backup files. You can also automate the downloads on a daily basis!

/mnt/c/Users/Lenovo/Downloads/ 

use this location path to download to Windows Machine